+++ src/index.ts
@@ -4,7 +4,13 @@ export { PinepostProvider, usePinepostLocale } from "./pinepost-provider";
 export type { PinepostProviderProps } from "./pinepost-provider";
 export { pinepostLocales } from "./locale";
 export type { PinepostLocale } from "./locale";
-export { pinepostThemes } from "./theme";
-export type { PinepostTheme } from "./theme";
+export {
+  createPinepostThemeCss,
+  mergePinepostThemeTokens,
+  pinepostThemePresets,
+  pinepostThemes,
+  pinepostThemeTokenNames
+} from "./theme";
+export type { PinepostTheme, PinepostThemeTokenName, PinepostThemeTokens } from "./theme";
 
 export * from "./components";